Update README.md
I originally didn't use this project because I thought it was a process monitor for the server, which seems orthogonal to HangFire.  I realized much later on revisiting this page that it is monitoring the CPU and RAM of just the Hangfire process, which is super useful.  Updated readme to clarify.